+++ FluidWarfare.Render.Vulkan/Camera/VulkanCameraMatrices.cs
@@ -93,7 +93,7 @@ public static class VulkanCameraMatrices
         [
             f / aspect, 0, 0, 0,
             0, -f, 0, 0,
-            0, 0, near / range, -1,
+            0, 0, far / range, -1,
             0, 0, near * far / range, 0
         ];
     }